pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "websockets>=12.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]
